+++ src/components/QuranReader/ContextMenu.tsx
@@ -1,13 +1,12 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 
 import classNames from 'classnames';
-import { shallowEqual, useDispatch, useSelector } from 'react-redux';
+import { shallowEqual, useSelector } from 'react-redux';
 
 import styles from './ContextMenu.module.scss';
 
-import useScrollDirection, { ScrollDirection } from 'src/hooks/useScrollDirection';
 import { selectNavbar } from 'src/redux/slices/navbar';
-import { selectContextMenu, setIsExpanded } from 'src/redux/slices/QuranReader/contextMenu';
+import { selectContextMenu } from 'src/redux/slices/QuranReader/contextMenu';
 import { selectNotes } from 'src/redux/slices/QuranReader/notes';
 import { selectLastReadVerseKey } from 'src/redux/slices/QuranReader/readingTracker';
 import { getChapterData, getChapterReadingProgress } from 'src/utils/chapter';
@@ -15,22 +14,10 @@ import { getJuzNumberByHizb } from 'src/utils/juz';
 import { getVerseNumberFromKey } from 'src/utils/verse';
 
 const ContextMenu = () => {
-  const dispatch = useDispatch();
   const isSideBarVisible = useSelector(selectNotes, shallowEqual).isVisible;
   const { isExpanded } = useSelector(selectContextMenu, shallowEqual);
   const isNavbarVisible = useSelector(selectNavbar, shallowEqual).isVisible;
   const { verseKey, chapterId, page, hizb } = useSelector(selectLastReadVerseKey, shallowEqual);
-  const onDirectionChange = useCallback(
-    (direction: ScrollDirection) => {
-      if (direction === ScrollDirection.Up && !isExpanded) {
-        dispatch({ type: setIsExpanded.type, payload: true });
-      } else if (direction === ScrollDirection.Down && isExpanded) {
-        dispatch({ type: setIsExpanded.type, payload: false });
-      }
-    },
-    [dispatch, isExpanded],
-  );
-  useScrollDirection(onDirectionChange);
   const chapterData = useMemo(() => {
     return chapterId ? getChapterData(chapterId) : null;
   }, [chapterId]);
